Refactor switchports. Handle edge cases
