+++ pyproject.toml
@@ -56,6 +56,7 @@ dependencies = [
   "django-sitetree==1.18.0",
   "django-taggit",
   "django-threadlocals==0.8",
+  "djangorestframework-stubs>=3.17.0",
   "djangorestframework==3.15.0",
   "djangorestframework_xml==2.0.0",
   "django~=4.2.0",
@@ -101,7 +102,7 @@ dependencies = [
   "stevedore",
   "tblib>=3.1.0",
   "typing==3.6.6",
-  "wrapt==1.10.6"
+  "wrapt==1.10.6",
 ]
 dynamic = [
   "authors",
