+++ src/ralph/switchports/rackconfig/__init__.py
@@ -1,0 +1,24 @@
+"""Rack switch-column domain: switch resolution, port labels and overrides."""
+
+from ralph.switchports.rackconfig.overrides import (
+    build_override_map,
+    effective_switch,
+    sync_override,
+)
+from ralph.switchports.rackconfig.port_labels import to_display, to_switch_label
+from ralph.switchports.rackconfig.switch_resolver import (
+    rack_data_center,
+    resolve_switch,
+    resolve_switch_by_position,
+)
+
+__all__ = [
+    "build_override_map",
+    "effective_switch",
+    "rack_data_center",
+    "resolve_switch",
+    "resolve_switch_by_position",
+    "sync_override",
+    "to_display",
+    "to_switch_label",
+]
