+++ src/ralph/switchports/views.py
@@ -1,78 +1,33 @@
 import logging
 
-from django import forms
 from django.contrib import messages
 from django.db import transaction
-from django.forms import inlineformset_factory
 from django.http import HttpResponseRedirect, JsonResponse
 from django.utils.functional import cached_property
 
 from ralph.admin.views.extra import RalphDetailView
-from ralph.data_center.models import DataCenterAsset
 from ralph.lib.external_services import InternalService
-from ralph.switchports import connections
+from ralph.switchports import grid
+from ralph.switchports.constants import (
+    grid_force_field,
+    grid_override_field,
+    grid_port_field,
+)
+from ralph.switchports.forms import (
+    RackConfigurationForm,
+    RackSwitchConfigurationFormSet,
+)
+from ralph.switchports.grid_edit import CellEdit, apply_cell_edit
 from ralph.switchports.models import (
     BackendValidationResult,
-    Port,
     RackConfiguration,
-    RackSwitchConfiguration,
-    RackSwitchConfigurationOverride,
     RefreshJobStatus,
     SwitchportRefreshJob,
-    ValidationStatus,
 )
-from ralph.switchports.presentation import build_validation_context
-from ralph.switchports.rack_configuration import _parse
-from ralph.switchports.validation import refresh_validation_for_rack  # noqa: F401
+from ralph.switchports.rackconfig.overrides import build_override_map
+from ralph.switchports.rackconfig.switch_resolver import rack_data_center
 
 logger = logging.getLogger(__name__)
-
-
-def _port_label_for_switch(port_number_str):
-    """Convert user-entered port number to switch port label.
-
-    If the value already contains '/', treat it as a full label.
-    Otherwise, format it as '0/0/{number}'.
-    """
-    port_number_str = port_number_str.strip()
-    if "/" in port_number_str:
-        return port_number_str
-    return f"0/0/{port_number_str}"
-
-
-def _resolve_switch(identifier, dc=None):
-    """Resolve a user-entered switch identifier.
-
-    Accepts a barcode, a hostname, or a ``rack{rack_number}u{position}``
-    position string resolved within ``dc`` the same way as
-    ``ralph.switchports.rack_configuration._parse`` (e.g. ``rack12u42`` is the
-    asset at position 42 of the rack whose name ends with ``12``). The position
-    format is only tried when a data center context (``dc``) is available.
-    Returns the matching DataCenterAsset or None.
-    """
-    identifier = identifier.strip()
-    if not identifier:
-        return None
-    try:
-        return DataCenterAsset.objects.get(barcode=identifier)
-    except DataCenterAsset.DoesNotExist:
-        pass
-    try:
-        return DataCenterAsset.objects.get(hostname=identifier)
-    except DataCenterAsset.DoesNotExist:
-        pass
-    if dc is not None:
-        try:
-            return _parse(identifier, dc)
-        except ValueError:
-            return None
-    return None
-
-
-def _rack_data_center(rack):
-    """Best-effort data center for a rack (may be None if not placed)."""
-    server_room = getattr(rack, "server_room", None)
-    return getattr(server_room, "data_center", None)
 
 
 class RackSwitchportGridView(RalphDetailView):
@@ -93,137 +48,6 @@ class RackSwitchportGridView(RalphDetailView):
         """
         rack_config, _ = RackConfiguration.objects.get_or_create(rack=self.object)
         return rack_config
-
-    def _get_rack_assets(self):
-        """Get all DataCenterAssets in this rack, ordered by position."""
-        return (
-            DataCenterAsset.objects.filter(rack=self.object)
-            .select_related("model", "model__category")
-            .order_by("-position", "slot_no", "hostname")
-        )
-
-    def _get_switch_configs(self):
-        """Get all RackSwitchConfigurations for this rack, ordered by label."""
-        return (
-            self.rack_configuration.switches.select_related("switch")
-            .order_by("label")
-        )
-
-    def _build_override_map(self, switch_configs):
-        """Map (asset_id, switch_config_id) -> override switch (DataCenterAsset)."""
-        sc_ids = [sc.id for sc in switch_configs]
-        override_map = {}
-        if not sc_ids:
-            return override_map
-        overrides = RackSwitchConfigurationOverride.objects.filter(
-            rack_switch_configuration_id__in=sc_ids
-        ).select_related("switch")
-        for override in overrides:
-            override_map[
-                (override.data_center_asset_id, override.rack_switch_configuration_id)
-            ] = override.switch
-        return override_map
-
-    def _effective_switch(self, sc, asset_id, override_map):
-        """The switch an asset connects to for a column (override or default)."""
-        return override_map.get((asset_id, sc.id)) or sc.switch
-
-    def _build_connection_map(self, assets, switch_configs):
-        """Build a map of (asset_id, switch_config_id) -> port info.
-
-        Matches an asset's port to a column by label, then records the switch
-        port it is connected to and which switch that actually is (which may be
-        an override switch rather than the column default).
-        """
-        label_to_sc = {sc.label: sc for sc in switch_configs}
-        asset_ids = [a.id for a in assets]
-
-        if not asset_ids or not label_to_sc:
-            return {}
-
-        asset_ports = (
-            Port.objects.filter(
-                data_center_asset_id__in=asset_ids,
-                label__in=list(label_to_sc.keys()),
-            )
-            .select_related(
-                "connectionmember__connection",
-            )
-            .prefetch_related(
-                "connectionmember__connection__members__port__data_center_asset",
-            )
-        )
-
-        connection_map = {}
-        for port in asset_ports:
-            sc = label_to_sc.get(port.label)
-            if sc is None:
-                continue
-            conn_member = getattr(port, "connectionmember", None)
-            if not conn_member:
-                continue
-            connection = conn_member.connection
-            for member in connection.members.all():
-                if member.port_id == port.id:
-                    continue
-                remote_port = member.port
-                remote_asset = remote_port.data_center_asset
-                # Only treat the remote as a switch cell if it is not another
-                # server in this rack's asset set (heuristic: it is a switch if
-                # it differs from the local asset).
-                if remote_asset.id == port.data_center_asset_id:
-                    continue
-                connection_map[(port.data_center_asset_id, sc.id)] = {
-                    "switch_port_label": remote_port.label,
-                    "asset_port_label": port.label,
-                    "actual_switch_id": remote_asset.id,
-                }
-        return connection_map
-
-    def _build_validation_map(self, switch_configs):
-        """Build a map of (switch_id, port_label) -> BackendValidationResult.
-
-        Also returns per-switch status (SWITCH_NOT_FOUND if the sentinel exists).
-        """
-        results = BackendValidationResult.objects.filter(
-            rack_configuration=self.rack_configuration,
-        ).select_related("remote_asset")
-
-        validation_map = {}
-        switch_status = {}
-        for result in results:
-            if result.port_label == "__switch__":
-                switch_status[result.switch_id] = result.status
-            else:
-                validation_map[(result.switch_id, result.port_label)] = result
-
-        return validation_map, switch_status
-
-    def _build_client_validation(self, switch_configs, validation_map, switch_status):
-        """Serialize default-switch validation for instant, client-side checks.
-
-        Only default (non-override) switches are included: the grid preloads
-        every validated port so typing a port number gives immediate feedback
-        without any AJAX. Overrides are intentionally excluded.
-        """
-        client = {}
-        for sc in switch_configs:
-            if not sc.backend_validation:
-                continue
-            ports = {}
-            for (switch_id, port_label), vr in validation_map.items():
-                if switch_id != sc.switch_id:
-                    continue
-                ports[port_label] = {
-                    "status": vr.status,
-                    "remote_asset_id": vr.remote_asset_id,
-                    "remote_hostname": vr.remote_hostname,
-                }
-            client[str(sc.id)] = {
-                "switch_not_found": sc.switch_id in switch_status,
-                "ports": ports,
-            }
-        return client
 
     def _latest_refresh_job(self):
         return (
@@ -257,11 +81,13 @@ class RackSwitchportGridView(RalphDetailView):
 
     def get_context_data(self, **kwargs):
         context = super().get_context_data(**kwargs)
-        assets = list(self._get_rack_assets())
-        switch_configs = list(self._get_switch_configs())
-        override_map = self._build_override_map(switch_configs)
-        connection_map = self._build_connection_map(assets, switch_configs)
-        validation_map, switch_status = self._build_validation_map(switch_configs)
+        assets = list(grid.get_rack_assets(self.object))
+        switch_configs = list(grid.get_switch_configs(self.rack_configuration))
+        override_map = build_override_map(switch_configs)
+        connection_map = grid.build_connection_map(assets, switch_configs)
+        validation_map, switch_status = grid.build_validation_map(
+            self.rack_configuration
+        )
 
         # Check if we have any validation results at all
         has_validation = bool(validation_map) or bool(switch_status)
@@ -270,76 +96,32 @@ class RackSwitchportGridView(RalphDetailView):
         last_refresh = None
         if has_validation:
             last_result = (
-                BackendValidationResult.objects.filter(rack_configuration=self.rack_configuration)
+                BackendValidationResult.objects.filter(
+                    rack_configuration=self.rack_configuration
+                )
                 .order_by("-modified")
                 .first()
             )
             if last_result:
                 last_refresh = last_result.modified
 
-        rows = []
-        for asset in assets:
-            cells = []
-            for sc in switch_configs:
-                effective_switch = self._effective_switch(sc, asset.id, override_map)
-                override_switch = override_map.get((asset.id, sc.id))
-                conn_info = connection_map.get((asset.id, sc.id))
-                if conn_info:
-                    switch_port_label = conn_info["switch_port_label"]
-                    display_value = switch_port_label
-                    if switch_port_label.startswith("0/0/"):
-                        display_value = switch_port_label[4:]
-                else:
-                    switch_port_label = None
-                    display_value = ""
-
-                # Validation for this cell (keyed by the effective switch)
-                validation = None
-                if not sc.backend_validation:
-                    # Backend validation disabled: no netmaker column.
-                    pass
-                elif switch_port_label and effective_switch.id in switch_status:
-                    validation = {
-                        "status": ValidationStatus.SWITCH_NOT_FOUND,
-                        "css_class": "validation-error",
-                        "label": "SWITCH N/F",
-                    }
-                elif switch_port_label:
-                    vr = validation_map.get((effective_switch.id, switch_port_label))
-                    if vr is not None:
-                        validation = build_validation_context(
-                            vr, expected_asset_id=asset.id
-                        )
-                    elif has_validation:
-                        validation = {
-                            "status": ValidationStatus.PORT_NOT_FOUND,
-                            "css_class": "validation-warning",
-                            "label": "PORT N/F",
-                        }
-
-                cells.append(
-                    {
-                        "switch_config": sc,
-                        "value": display_value,
-                        "field_name": f"port_{asset.id}_{sc.id}",
-                        "override_field_name": f"override_{asset.id}_{sc.id}",
-                        "override_value": (
-                            override_switch.barcode or override_switch.hostname
-                            if override_switch
-                            else ""
-                        ),
-                        "is_override": bool(override_switch),
-                        "validation": validation,
-                    }
-                )
-            rows.append({"asset": asset, "cells": cells})
-
         context["switch_configs"] = switch_configs
         context["switch_status"] = switch_status
-        context["rows"] = rows
+        conflicts = getattr(self, "_conflicts", {})
+        context["rows"] = grid.build_grid_rows(
+            assets,
+            switch_configs,
+            override_map,
+            connection_map,
+            validation_map,
+            switch_status,
+            has_validation,
+            conflicts=conflicts,
+        )
+        context["has_conflicts"] = bool(conflicts)
         context["has_validation"] = has_validation
         context["last_refresh"] = last_refresh
-        context["client_validation"] = self._build_client_validation(
+        context["client_validation"] = grid.build_client_validation(
             switch_configs, validation_map, switch_status
         )
         context["refresh_job"] = self._latest_refresh_job()
@@ -350,110 +132,64 @@ class RackSwitchportGridView(RalphDetailView):
         if "refresh_validation" in request.POST:
             return self._handle_refresh(request)
 
-        # Handle save connections
-        switch_configs = list(self._get_switch_configs())
-        assets = list(self._get_rack_assets())
-        connection_map = self._build_connection_map(assets, switch_configs)
+        return self._save_connections(request)
+
+    def _save_connections(self, request):
+        """Apply the grid edits, deferring any conflicting port steals.
+
+        Non-conflicting cells are applied immediately. If any cell would steal a
+        switch port already used by another asset (and was not explicitly
+        confirmed), the whole grid is re-rendered with those cells highlighted
+        and an "overwrite" checkbox, so the user decides per conflict.
+        """
+        switch_configs = list(grid.get_switch_configs(self.rack_configuration))
+        assets = list(grid.get_rack_assets(self.object))
+        connection_map = grid.build_connection_map(assets, switch_configs)
+        dc = rack_data_center(self.object)
 
         created_count = 0
         removed_count = 0
         errors = []
-        dc = _rack_data_center(self.object)
+        conflicts = {}
 
         with transaction.atomic():
             for asset in assets:
                 for sc in switch_configs:
-                    field_name = f"port_{asset.id}_{sc.id}"
-                    override_field = f"override_{asset.id}_{sc.id}"
-                    new_value = request.POST.get(field_name, "").strip()
-                    override_value = request.POST.get(override_field, "").strip()
-
-                    # Resolve the target switch (override or column default).
-                    if override_value:
-                        target_switch = _resolve_switch(override_value, dc)
-                        if target_switch is None:
-                            errors.append(
-                                f"{asset.hostname} / {sc.label}: "
-                                f"unknown switch '{override_value}'"
-                            )
-                            continue
-                    else:
-                        target_switch = sc.switch
-
-                    self._sync_override(sc, asset, target_switch, override_value)
-
-                    existing = connection_map.get((asset.id, sc.id))
-                    existing_label = existing["switch_port_label"] if existing else None
-                    existing_switch_id = (
-                        existing["actual_switch_id"] if existing else None
+                    edit = CellEdit(
+                        new_value=request.POST.get(
+                            grid_port_field(asset.id, sc.id), ""
+                        ),
+                        override_value=request.POST.get(
+                            grid_override_field(asset.id, sc.id), ""
+                        ),
+                        forced=grid_force_field(asset.id, sc.id) in request.POST,
                     )
-
-                    if new_value:
-                        new_switch_port_label = _port_label_for_switch(new_value)
-                    else:
-                        new_switch_port_label = None
-
-                    unchanged = (
-                        existing_label == new_switch_port_label
-                        and existing_switch_id == (
-                            target_switch.id if new_switch_port_label else None
-                        )
-                    )
-                    if unchanged:
-                        continue
-
-                    if existing and not new_switch_port_label:
-                        # Disconnect
-                        try:
-                            asset_port = Port.objects.get(
-                                label=existing["asset_port_label"],
-                                data_center_asset=asset,
-                            )
-                            connections.disconnect(asset_port)
-                            removed_count += 1
-                        except Port.DoesNotExist:
-                            pass
-                        continue
-
-                    if new_switch_port_label:
-                        # Connect (or reconnect, possibly to a different switch)
-                        try:
-                            asset_port, _ = Port.objects.get_or_create(
-                                label=sc.label,
-                                data_center_asset=asset,
-                            )
-                            switch_port, _ = Port.objects.get_or_create(
-                                label=new_switch_port_label,
-                                data_center_asset=target_switch,
-                            )
-                            connections.connect(asset_port, switch_port)
-                            created_count += 1
-                        except Exception as e:
-                            errors.append(f"{asset.hostname} / {sc.label}: {e}")
+                    result = apply_cell_edit(asset, sc, edit, connection_map, dc)
+                    created_count += result.created
+                    removed_count += result.removed
+                    if result.error:
+                        errors.append(result.error)
+                    if result.conflict is not None:
+                        conflicts[(asset.id, sc.id)] = result.conflict
 
         if created_count:
             messages.success(request, f"Created/updated {created_count} connection(s).")
         if removed_count:
             messages.success(request, f"Removed {removed_count} connection(s).")
-        if errors:
-            for error in errors:
-                messages.error(request, error)
+        for error in errors:
+            messages.error(request, error)
+
+        if conflicts:
+            messages.warning(
+                request,
+                f"{len(conflicts)} switch port(s) are already in use. Tick "
+                "'overwrite' on the highlighted cells and save again to "
+                "reassign them.",
+            )
+            self._conflicts = conflicts
+            return self.render_to_response(self.get_context_data())
 
         return HttpResponseRedirect(request.path)
-
-    def _sync_override(self, sc, asset, target_switch, override_value):
-        """Create, update or drop the per-server override for a cell."""
-        if override_value and target_switch.id != sc.switch_id:
-            RackSwitchConfigurationOverride.objects.update_or_create(
-                rack_switch_configuration=sc,
-                data_center_asset=asset,
-                defaults={"switch": target_switch},
-            )
-        else:
-            RackSwitchConfigurationOverride.objects.filter(
-                rack_switch_configuration=sc,
-                data_center_asset=asset,
-            ).delete()
 
     def _handle_refresh(self, request):
         """Enqueue an asynchronous netmaker refresh for the whole rack."""
@@ -484,77 +220,6 @@ class RackSwitchportGridView(RalphDetailView):
             messages.error(request, f"Failed to start refresh: {e}")
 
         return HttpResponseRedirect(request.path)
-
-
-class RackConfigurationForm(forms.ModelForm):
-    class Meta:
-        model = RackConfiguration
-        fields = ["description"]
-        widgets = {
-            "description": forms.Textarea(attrs={"rows": 3}),
-        }
-
-
-class RackSwitchConfigurationForm(forms.ModelForm):
-    """Row form for a single switch column of a rack.
-
-    The ``switch`` FK is entered as a free-text barcode/hostname (resolved the
-    same way as the switchport grid) instead of a giant asset dropdown.
-    """
-
-    switch_identifier = forms.CharField(
-        label="Switch (barcode, hostname or rackNuM)",
-        required=False,
-        help_text=(
-            "Barcode, hostname, or a rack position like "
-            "<code>rack12u42</code> (switch at position 42 of the rack whose "
-            "name ends with 12, in this rack's data center)."
-        ),
-    )
-
-    class Meta:
-        model = RackSwitchConfiguration
-        fields = ["label", "backend_validation"]
-
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-        if self.instance and self.instance.pk and self.instance.switch_id:
-            switch = self.instance.switch
-            self.fields["switch_identifier"].initial = (
-                switch.barcode or switch.hostname
-            )
-
-    def clean(self):
-        cleaned_data = super().clean()
-        if cleaned_data.get("DELETE"):
-            return cleaned_data
-        label = (cleaned_data.get("label") or "").strip()
-        identifier = (cleaned_data.get("switch_identifier") or "").strip()
-        # Untouched blank extra row - nothing to validate or save.
-        if not label and not identifier:
-            return cleaned_data
-        rack_config = getattr(self.instance, "rack_configuration", None)
-        dc = _rack_data_center(getattr(rack_config, "rack", None))
-        switch = _resolve_switch(identifier, dc)
-        if switch is None:
-            self.add_error(
-                "switch_identifier",
-                "Unknown switch '{}'.".format(identifier)
-                if identifier
-                else "This field is required.",
-            )
-        else:
-            self.instance.switch = switch
-        return cleaned_data
-
-
-RackSwitchConfigurationFormSet = inlineformset_factory(
-    RackConfiguration,
-    RackSwitchConfiguration,
-    form=RackSwitchConfigurationForm,
-    extra=1,
-    can_delete=True,
-)
 
 
 class RackConfigurationView(RalphDetailView):
